requirements.txt
# Core dependencies
click>=8.0.0
pyyaml>=6.0
kubernetes>=24.0.0
requests>=2.28.0

# AWS Bedrock integration
boto3>=1.26.0
botocore>=1.29.0

# Data handling
dataclasses-json>=0.5.7
pydantic>=1.10.0

# Development dependencies
pytest>=7.0.0
pytest-cov>=4.0.0
black>=22.0.0
flake8>=5.0.0
mypy>=1.0.0

# Optional dependencies for enhanced functionality
rich>=12.0.0  # For better CLI output
tqdm>=4.64.0  # For progress bars